requirements.txt
flask
python-dotenv
langchain
langchain-community
langchain-google-genai
tavily-python
